pyproject.toml
[project]
name = "brinksmanship"
version = "0.1.0"
description = "A game-theoretic strategy simulation exploring nuclear brinkmanship and coordination games"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["game-theory", "strategy", "simulation", "claude", "ai", "brinkmanship"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "claude-agent-sdk>=0.1.0",
    "simple-term-menu>=1.6.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "argon2-cffi>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]
webapp = [
    "flask>=3.0.0",
    "flask-login>=0.6.0",
    "flask-sqlalchemy>=3.1.0",
    "markdown>=3.5.0",
]

[project.scripts]
brinksmanship = "brinksmanship.cli.app:main"
brinksmanship-web = "brinksmanship.webapp.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brinksmanship"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4"]

[tool.ruff.lint.per-file-ignores]
# Prompt templates have long lines by design
"src/brinksmanship/prompts.py" = ["E501"]
# Scripts have long print statements and game theory variable names (C, D)
"scripts/*" = ["E501", "N806"]
# Test files check import availability with try/except
"test_*.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "llm_integration: marks tests requiring LLM API calls",
    "e2e: marks end-to-end Playwright tests",
    "webapp: marks webapp-specific tests",
]

[dependency-groups]
dev = [
    "cairosvg>=2.8.2",
]
